requirements.txt
randomname==0.2.1
Pillow==10.2.0
piexif==1.1.3
python-dotenv==1.0.1 
requests==2.28.1